Prints Out Solution Path + New Memory Saving Technique
When a solution is reached it prints out a complete path to that
solution.

Now the program does not run out of memory when solving complex
puzzles. This is because instead of storing a full puzzleBoard in each
boardNode  it only stores a Long representation of the puzzleBoard.

Even though the program does not run out of memory, it still takes too
long to solve complex puzzles.
I think the next step should be to find an algorithm that solves the
puzzle in a reasonable time frame.